Extract SafeFetch32/N to stubRoutines.inline.hpp
